Added IdentityMapper class. Refactored privilege table to it.
The GameMaster should ensure that only registered characters are recognized
by IdentityMap. This prevents malicious Villager implementations from creating
an instance of another villager and passing it to the GameMaster as their
identity in order to gain a privilege for an action they have no right to.

This would, however, assume that the name + rolestring combination is unique.